requirements.txt
Django==4.2.16
djangorestframework==3.16.1
djangorestframework-simplejwt==5.3.0
mongoengine
pymongo==4.15.3
django-cors-headers==4.9.0
python-decouple==3.8
pytz==2024.1
razorpay==1.4.2
gunicorn==21.2.0
